+++ backend/src/main/java/com/mysite/sbb/user/UserController.java
@@ -4,6 +4,10 @@ import com.mysite.sbb.answer.AnswerDto;
 import com.mysite.sbb.comment.CommentDto;
 import com.mysite.sbb.global.ResponseDto;
 import com.mysite.sbb.question.QuestionDto;
+import com.mysite.sbb.user.dto.EmailRequestDto;
+import com.mysite.sbb.user.dto.PasswordRequestDto;
+import com.mysite.sbb.user.dto.UserRequestDto;
+import com.mysite.sbb.user.dto.UserResponseDto;
 import com.mysite.sbb.util.DataNotFoundException;
 import com.mysite.sbb.answer.Answer;
 import com.mysite.sbb.answer.AnswerService;
