+++ backend/src/main/java/com/mysite/sbb/user/dto/EmailRequestDto.java
@@ -1,4 +1,4 @@
-package com.mysite.sbb.user;
+package com.mysite.sbb.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserResponseDto {
-    private String username;
+public class EmailRequestDto {
     private String email;
 }
