+++ backend/src/main/java/com/mysite/sbb/user/dto/PasswordRequestDto.java
@@ -1,4 +1,4 @@
-package com.mysite.sbb.user;
+package com.mysite.sbb.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class EmailRequestDto {
-    private String email;
+public class PasswordRequestDto {
+    private String currentPassword;
+    private String newPassword;
 }
