+++ backend/src/main/java/com/mysite/sbb/user/dto/UserResponseDto.java
@@ -1,4 +1,4 @@
-package com.mysite.sbb.user;
+package com.mysite.sbb.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class PasswordRequestDto {
-    private String currentPassword;
-    private String newPassword;
+public class UserResponseDto {
+    private String username;
+    private String email;
 }
